Address PR feedback. Consolidate transport error formatting
