+++ elixir/test/portal/okta/error_codes_test.exs
@@ -227,73 +227,4 @@ defmodule Portal.Okta.ErrorCodesTest do
     end
   end
 
-  describe "format_transport_error/1" do
-    test "handles DNS lookup failure (nxdomain)" do
-      error = %Req.TransportError{reason: :nxdomain}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "DNS lookup failed"
-      assert result =~ "domain is spelled correctly"
-    end
-
-    test "handles timeout" do
-      error = %Req.TransportError{reason: :timeout}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "timed out"
-      assert result =~ "network connectivity"
-    end
-
-    test "handles connect_timeout" do
-      error = %Req.TransportError{reason: :connect_timeout}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "timed out"
-    end
-
-    test "handles connection refused" do
-      error = %Req.TransportError{reason: :econnrefused}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "Connection refused"
-      assert result =~ "domain is correct"
-    end
-
-    test "handles connection closed" do
-      error = %Req.TransportError{reason: :closed}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "closed unexpectedly"
-    end
-
-    test "handles TLS alert" do
-      error = %Req.TransportError{reason: {:tls_alert, {:certificate_expired, "test"}}}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "TLS error"
-      assert result =~ "certificate_expired"
-    end
-
-    test "handles host unreachable" do
-      error = %Req.TransportError{reason: :ehostunreach}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "Host is unreachable"
-    end
-
-    test "handles network unreachable" do
-      error = %Req.TransportError{reason: :enetunreach}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "Network is unreachable"
-    end
-
-    test "handles unknown transport errors" do
-      error = %Req.TransportError{reason: :some_unknown_error}
-      result = ErrorCodes.format_transport_error(error)
-
-      assert result =~ "Network error"
-      assert result =~ "some_unknown_error"
-    end
-  end
 end
